clean up code for identify functions
